+++ src/angular/hq/src/app/charge-code/charge-code.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+  selector: 'hq-charge-code',
+  standalone: true,
+  imports: [RouterOutlet],
+  templateUrl: './charge-code.component.html'
+})
+export class ChargeCodeComponent {
+
+}