+++ src/angular/hq/src/app/psr/psrlist/services/pstlistService.ts
@@ -1,0 +1,114 @@
+import { FormControl } from "@angular/forms";
+import { BehaviorSubject } from "rxjs/internal/BehaviorSubject";
+import { ProjectStatus } from "../../../clients/client-details.service";
+import { GetPSRTimeRecordStaffV1 } from "../../../models/PSR/get-psr-time-v1";
+import { HQService } from "../../../services/hq.service";
+import { Injectable } from "@angular/core";
+
+
+export enum ActivityName {
+  Support = 0,
+  Development = 1,
+  Todo = 2,
+}
+
+@Injectable({
+  providedIn: 'root',
+})
+export class PsrListService {
+  staffMembers$ = new BehaviorSubject<GetPSRTimeRecordStaffV1[]>([]);
+  search = new FormControl<string | null>('');
+  roaster = new FormControl<string | null>('');
+
+  projectStatus = new FormControl<ProjectStatus>(ProjectStatus.InProduction);
+  activityName = new FormControl<ActivityName>(ActivityName.Development);
+  staffMember = new FormControl<string | null>(null);
+  isSubmitted = new FormControl<boolean | null>(null);
+  startDate = new FormControl<Date | null>(null);
+  endDate = new FormControl<Date | null>(null);
+
+  // Pagination form controls
+  itemsPerPage = new FormControl(20, { nonNullable: true });
+  page = new FormControl<number>(1, { nonNullable: true });
+
+  ProjectStatus = ProjectStatus;
+  ActivityName = ActivityName;
+
+  showProjectStatus$ = new BehaviorSubject<boolean>(true);
+  showSearch$ = new BehaviorSubject<boolean>(true);
+  showStaffMembers$ = new BehaviorSubject<boolean>(true);
+  showIsSubmitted$ = new BehaviorSubject<boolean>(true);
+  showStartDate$ = new BehaviorSubject<boolean>(true);
+  showEndDate$ = new BehaviorSubject<boolean>(true);
+
+
+
+
+  showActivityName$ = new BehaviorSubject<boolean>(true);
+  showRoaster$ = new BehaviorSubject<boolean>(true);
+
+  constructor(private hqService: HQService) {}
+
+  resetFilter() {
+    this.search.setValue('');
+    this.projectStatus.setValue(ProjectStatus.InProduction);
+    this.activityName.setValue(ActivityName.Development);
+    this.staffMember.setValue(null);
+    this.roaster.setValue('');
+    this.isSubmitted.setValue(null);
+    this.startDate.setValue(null);
+    this.endDate.setValue(null);
+  }
+  showProjectStatus() {
+    this.showProjectStatus$.next(true);
+  }
+  showSearch() {
+    this.showSearch$.next(true);
+  }
+
+  hideSearch() {
+    this.showSearch$.next(false);
+  }
+  showStaffMembers() {
+    this.showStaffMembers$.next(true);
+  }
+  hideStaffMembers() {
+    this.showStaffMembers$.next(false);
+  }
+  hideProjectStatus() {
+    this.showProjectStatus$.next(false);
+  }
+
+  showActivityName() {
+    this.showActivityName$.next(true);
+  }
+  hideActivityName() {
+    this.showActivityName$.next(false);
+  }
+  showRoaster() {
+    this.showRoaster$.next(true);
+  }
+  hideRoaster() {
+    this.showRoaster$.next(false);
+  }
+  showIsSubmitted() {
+    this.showIsSubmitted$.next(true);
+  }
+  hideIsSubmitted() {
+    this.showIsSubmitted$.next(false);
+  }
+
+  showStartDate() {
+    this.showStartDate$.next(true);
+  }
+  hideStartDate() {
+    this.showStartDate$.next(false);
+  }
+  showEndDate() {
+    this.showEndDate$.next(true);
+  }
+  hideEndDate() {
+    this.showEndDate$.next(false);
+  }
+}
+