+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Modal service with custom alert, confirm and prompt modals
+
+## [0.3.0] - 2024-06-09
+
+### Fixed
+- Page titles on staff routes
+- Client modal pagination size too large
+
+### Changed
+- Preserve filters, page size and current page when navigating away from PSR list into PSR time and report view
+
+### Added
+- "No matching records found" when no results are returned from API
+- Saving indicator to PSR report when saving markdown to API
+- WIP project management
+- WIP charge code management
+
 ## [0.2.1] - 2024-06-05
 
 ### Added
@@ -29,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[unreleased]: https://github.com/sancsoft/hq/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/sancsoft/hq/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/sancsoft/hq/releases/tag/v0.3.0
 [0.2.1]: https://github.com/sancsoft/hq/releases/tag/v0.2.1
 [0.2.0]: https://github.com/sancsoft/hq/releases/tag/v0.2.0
 [0.1.1]: https://github.com/sancsoft/hq/compare/v0.1.0...v0.1.1